requirements.txt
# Core ML/AI
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LangChain/LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-aws>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AWS CDK
aws-cdk-lib>=2.100.0
constructs>=10.3.0

# Data Processing
pydantic>=2.5.0
python-dateutil>=2.8.2
faker>=20.1.0

# API/Web
fastapi>=0.104.0
uvicorn>=0.24.0

# Monitoring/Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
moto>=4.2.0

# Development
black>=23.11.0
mypy>=1.7.0
flake8>=6.1.0